+++ siteapp/settings_application.py
@@ -15,6 +15,7 @@ INSTALLED_APPS += [
     'guidedmodules',
     'discussion',
     'controls',
+    'inventory',
     'system_settings',
 
     'loadtesting',